fix: 4 cargo errors only Windows-MSVC could catch
1. services/twitch.rs:234 — viewer_count was u64 from the Helix
   wire model, u32 in the public TwitchStream type. Cast at the
   boundary; viewer counts realistically fit in u32 (max ~4 B).
2. services/rawg.rs:169 — same shape: get_u64("id") into a u32
   field. Cast.
3. library/mod.rs (×2) — used AppError::Internal which doesn't
   exist in this enum. Switched to AppError::Other (already a
   free-form variant).
4. lib.rs:97 — menu_on_left_click is deprecated in tauri 2.11;
   renamed to show_menu_on_left_click.

These were all writeable-without-running-cargo bugs and slipped
through because Tauri's Linux build deps (gdk-sys, libdbus-sys)
need apt installs I don't have on this WSL — `cargo check` bombed
before pixiis itself was checked. Sorry about the iteration cost.

+++ src-tauri/src/services/rawg.rs
@@ -166,7 +166,7 @@ fn parse_game(obj: Value) -> RawgGameData {
         .unwrap_or_else(|| get_str("description"));
 
     RawgGameData {
-        id: get_u64("id"),
+        id: get_u64("id") as u32,
         name: get_str("name"),
         description,
         rating: get_f64("rating") as f32,
